Major Update #1
In this commit, the source code was further refined, with syntax and file placement being shuffled around and made better.

The major update includes code which makes annyang send recognized commands to the form, which can be then confirmed by the user. Essentially, the first and major barrier of making the software practical has been crosses.

It is now possible to turn a light on and off through voice commands.

The script, arduino and other relevant files have been updated.

Dynamic usertesting will henceforth commence, to see the pros, cons and various possible improvements on the current prototype and module in hopes of creating an improved second iteration.