Major Update #2
In this commit, the source code was further refined, with syntax and file placement being shuffled around and made better. The improvements, and some new additions, were added in accordance to the results from the first user testing session. This is, therefore, iteration #2. Iteration 1 is archived locally, and will be sorted out, and uploaded locally for archival purposes.

The major update includes code which allows annyang to recognize named lights. For the purpose of this prototype, the existing named lights are assumed to be "a", "b" and "c" . These lights correspond to external light diodes. On command recognition, the user input through the command is sent through various validation forms which confirm if the user input is relevant to the expected, and possible, results.

It is now possible to, for example, say the phrases "turn light A off" or "turn light C on"

The start of other minor functions, such as scrolling through lights, are under consideration. For now, the major new function is within functioning scope.