Major Update #3
In this commit, the source code was further refined, with syntax and file placement being shuffled around and made better. Various errors, on both the javascript and arduino part, were patched.

In this update, iteration #2 was finalized. Iteration 1 is is now available on github as well.

The major update includes code which allows annyang to recognize named lights without errors.

The start of other minor functions, such as scrolling through lights, which were under construction, were scrapped for reasons concerning priority and time. For now, the major new function is within functioning scope without error.

 /Jovan